Statistic calcultaons moved to CLI.

+++ Core/OHS_program_api.Application/ViewModels/Safety/AccidentStatistic/VM_List_AccidentStatistic.cs
@@ -16,6 +16,6 @@
         public int? WorkingHoursUnderground { get; set; }
         public int? WorkingHoursSummary { get; set; }
         public int? LostDayOfWorkSummary { get; set; }
-        public int? AccidentSeverityRate { get; set; }
+        public double? AccidentSeverityRate { get; set; }
     }
 }
